require in our jswonwebtoken library we added when we init'd the project.
use the returned user._id value from mongo/mongoose to set the userId in the token. We will change the secret in the future, this was just a test to get back a valid jwt and it works. verifed via jwt.io after it was returned via Postman and sure enough there is ou user._id embedded in.