create a new middleware that is required for any of our interactions with the endpoints. This middleware: requireAuth will be an exported module available to the rest of the app.
It will take in the request and verify that the request has the necessary authorization header in it. If it does not we will immediately return a 401 and error that you must be logged in.

If there authorization header is present, we will then remove out the Bearer portion of the string, and just save our token. Then we can verify that token against our secret Key, which does need to be secured. If there is an error - we just return 401 must be logged in. If that passes the jwt.verify step, then we pull out the userId from the payload and then check in Mongo for that user and add that user to the request and call next. At that point the api we wrapped in the requireAuth would be called.