to do the salting and hashing we will use the bcrypt library. So require it.
then add a pre-save hook on the userSchema. Note that instead of using fat arrow function we use the keyword function. this has to do with the 'this' keyword and its binding to the outter User.js object vs the user itself.

add a pre-save hook that will check first if the password has been modified if no - just return to our next step. If it has been modified we need to salt the password via the genSalt with a complexity of 10. Then in the callback of the genSalt we can use the salt to call the hash method. Pass in the users password, the returned salt. then set the user.password (in Mongo) to the actual hash.

Then when we want to compare the input password from the user vs the stored salted hash in mongo, we have to apply the same to the input. So pass in the user input password, candidatePassword. Then return a new Promise that can get rejected with an error or if the password doesn't match. Essentially you will take in the candidatePassword and compare the password stored in mongo and then if all is correct resolver true on the promise.