adding in a new /signin route. In it we destructure out the email,passwod from the request.body.
Give some generic warnings if therre is no email or password provided. Give some generic warnings if there is no user matching the passed in email.

If there is an email,password supplied and the user is valid in the db, then we try/catch the process of calling our comparePassword function on the userSchema. As long as you pass the comparePassword function then we will create a new valid jwt with the user._id and our standard secret then send that back to the client. If the password comparison fails then generic message about invalid password or email.

We make those warning generic. You don't want a malicious user to figure out whether an email is valid or not.