Add a new schema for the "Track" and tie that mongoose Model to the new trackSchema.
Track will have userId, name, locations. Note that locations needs its own schema defined and available prior to using pointSchema for locations array. So define the pointSchema above that has a timestamp and coords object.

Do not need to map the pointSchema to mongo via the mongoose model because we only will have a Tracks model in the db and that will have the pointsSchema inside it.