creating a new route for the express app so we can get /tracks so create the trackRoutes.js file. RequireAuth for interaction with the route. Tell the new router to user the requireAuth.
use the .find method on the model so Track.find and pass in the userId to search tracks for. you can use the passed in req objec since it will have ther user attached after we go through the requireAuth middleware.

Export the /tracks router to the rest of the express app.

Then to make the /tracks route available import now to the index.js and ell the app to .use it .use(trackRoutes).